refactor: Organize templates into scalable category-based structure (#84)
Refactored template organization from a single 600-line file into a modular,
category-based structure to support scaling to 50-100+ templates.

## Changes

### New Directory Structure
```
src/terragrunt/templates/
├── index.ts                    # TemplatesManager (simplified from 606 to 115 lines)
├── registry.ts                 # Aggregates all templates
├── categories/
│   ├── backends.ts            # 3 backend templates (AWS S3, Azure Blob, GCP GCS)
│   ├── providers.ts           # 1 provider template (AWS)
│   ├── hooks.ts               # 2 hook templates (before, after)
│   ├── dependencies.ts        # 1 dependency template
│   ├── configuration.ts       # 2 config templates (inputs, version)
│   └── index.ts               # Exports all categories
└── loaders/
    └── builtin.ts             # Builtin template loader
```

### Benefits
- ✅ **Scalability**: Can easily handle 50-100+ templates (10-20 per category)
- ✅ **Organization**: Clear separation by functionality
- ✅ **Maintainability**: Each category file ~150 lines vs 600-line monolith
- ✅ **Type Safety**: Full TypeScript benefits maintained
- ✅ **Future-Proof**: Foundation for custom user templates
- ✅ **Non-Breaking**: Public API unchanged, all 532 tests pass

### Technical Details
- Extracted 9 templates into 5 category files
- Created registry system for template aggregation
- Simplified TemplatesManager from 606 to 115 lines (80% reduction)
- Updated imports in 5 files (src + tests)
- All existing functionality preserved

## Testing
- ✅ All 532 tests passing
- ✅ TypeScript compilation successful
- ✅ No breaking changes to public API
- ✅ Template loading verified
- ✅ Metadata generation works

## Related
Resolves #84
Part of Epic #9
