+++ src/terragrunt/templates/categories/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Template categories index
+ * Exports all template categories
+ */
+
+export { backendTemplates } from './backends.js';
+export { providerTemplates } from './providers.js';
+export { hookTemplates } from './hooks.js';
+export { dependencyTemplates } from './dependencies.js';
+export { configurationTemplates } from './configuration.js';
