+++ src/terragrunt/templates/loaders/builtin.ts
@@ -1,0 +1,14 @@
+/**
+ * Builtin Template Loader
+ * Loads templates defined in TypeScript category files
+ */
+
+import { ConfigTemplate } from '../../../types/templates.js';
+import { allTemplates } from '../registry.js';
+
+/**
+ * Load all built-in templates from the registry
+ */
+export function loadBuiltinTemplates(): ConfigTemplate[] {
+  return allTemplates;
+}
